feat: prepare rbac basic interface

+++ pkg/api/api_clusters.go
@@ -8,6 +8,10 @@ import (
 )
 
 func (api *ApiImpl) CreateCluster(c echo.Context) error {
+	if !api.rbac.IsAllowed("TODO username", "create", "cluster") {
+		return c.JSON(http.StatusForbidden, nil)
+	}
+
 	_, err := api.clusterManager.Create(nil)
 	if err != nil {
 		return c.JSON(http.StatusInternalServerError, nil)
@@ -17,6 +21,10 @@ func (api *ApiImpl) CreateCluster(c echo.Context) error {
 }
 
 func (api *ApiImpl) DeleteCluster(c echo.Context, id string) error {
+	if !api.rbac.IsAllowed("TODO username", "delete", "cluster") {
+		return c.JSON(http.StatusForbidden, nil)
+	}
+
 	if err := api.clusterManager.Delete(id); err != nil {
 		if errors.IsNotFound(err) {
 			return c.JSON(http.StatusNotFound, nil)
@@ -28,6 +36,10 @@ func (api *ApiImpl) DeleteCluster(c echo.Context, id string) error {
 }
 
 func (api *ApiImpl) GetCluster(c echo.Context, id string) error {
+	if !api.rbac.IsAllowed("TODO username", "get", "cluster") {
+		return c.JSON(http.StatusForbidden, nil)
+	}
+
 	_, err := api.clusterManager.Get(id)
 	if err != nil {
 		if errors.IsNotFound(err) {
@@ -40,6 +52,10 @@ func (api *ApiImpl) GetCluster(c echo.Context, id string) error {
 }
 
 func (api *ApiImpl) ListClusters(c echo.Context) error {
+	if !api.rbac.IsAllowed("TODO username", "list", "cluster") {
+		return c.JSON(http.StatusForbidden, nil)
+	}
+
 	_, err := api.clusterManager.List()
 	if err != nil {
 		return c.JSON(http.StatusInternalServerError, nil)
